thumbnail_service/requirements.txt
ffmpeg-python==0.2.0
boto3==1.35.24
celery==5.4.0
redis==5.0.8